flixel core classes fixes

+++ lib/AdamAtomic-flixel-8989e50/org/flixel/system/FlxQuadTree.as
@@ -258,13 +258,17 @@ package org.flixel.system
 		 */
 		public function destroy():void
 		{
-			_headA.destroy();
+			if(_headA != null)
+				_headA.destroy();
 			_headA = null;
-			_tailA.destroy();
+			if(_tailA != null)
+				_tailA.destroy();
 			_tailA = null;
-			_headB.destroy();
+			if(_headB != null)
+				_headB.destroy();
 			_headB = null;
-			_tailB.destroy();
+			if(_tailB != null)
+				_tailB.destroy();
 			_tailB = null;
 
 			if(_northWestTree != null)
@@ -564,9 +568,11 @@ package org.flixel.system
 				{
 					//Execute callback functions if they exist
 					if((_processingCallback == null) || _processingCallback(_object,checkObject))
+					{
 						overlapProcessed = true;
-					if(overlapProcessed && (_notifyCallback != null))
-						_notifyCallback(_object,checkObject);
+						if(_notifyCallback != null)
+							_notifyCallback(_object, checkObject);
+					}
 				}
 				_iterator = _iterator.next;
 			}
